feat: added progress bar option
