+++ Cargo.toml
@@ -28,8 +28,11 @@ log = "0.4.22"
 filetime = "0.2.25"
 symlink = "0.1.0"
 ffmpeg-next = { version = "7.1.0", optional = true }
-fern = { version = "0.7.0", features = ["colored"] }
+fern = { version = "0.7.0", features = ["colored"], optional = true }
+indicatif = { version = "0.17.8", optional = true }
+indicatif-log-bridge = { version = "0.2.3", optional = true }
 
 [features]
-default = []
-video = ["dep:ffmpeg-next"]
+default = ["binary"]
+video = ["dep:ffmpeg-next"] # enable video support
+binary = ["dep:fern", "dep:indicatif", "dep:indicatif-log-bridge"] # enable main.rs dependencies
